requirements.txt
discord.py>=2.5.2
aiohttp>=3.12.14
requests>=2.32.4
python-dotenv>=1.0.0
